+++ lib/schema_ferry/warnings.rb
@@ -1,8 +1,6 @@
 # frozen_string_literal: true
 
 module SchemaFerry
-  # Mix in (include for classes, extend for module_function modules) to emit
-  # "[schema_ferry]"-prefixed warnings to stderr.
   module Warnings
     private